+++ project/models.go
@@ -4,13 +4,13 @@ import "time"
 
 // Project represents a Pushpad project.
 type Project struct {
-	ID                           int        `json:"id,omitempty"`
-	SenderID                     int        `json:"sender_id,omitempty"`
+	ID                           int64      `json:"id,omitempty"`
+	SenderID                     int64      `json:"sender_id,omitempty"`
 	Name                         string     `json:"name,omitempty"`
 	Website                      string     `json:"website,omitempty"`
 	IconURL                      string     `json:"icon_url,omitempty"`
 	BadgeURL                     string     `json:"badge_url,omitempty"`
-	NotificationsTTL             *int       `json:"notifications_ttl,omitempty"`
+	NotificationsTTL             *int64     `json:"notifications_ttl,omitempty"`
 	NotificationsRequireInteract *bool      `json:"notifications_require_interaction,omitempty"`
 	NotificationsSilent          *bool      `json:"notifications_silent,omitempty"`
 	CreatedAt                    *time.Time `json:"created_at,omitempty"`
@@ -18,12 +18,12 @@ type Project struct {
 
 // ProjectCreateParams is the payload to create a project.
 type ProjectCreateParams struct {
-	SenderID                     *int    `json:"sender_id"`
+	SenderID                     *int64  `json:"sender_id"`
 	Name                         *string `json:"name"`
 	Website                      *string `json:"website"`
 	IconURL                      *string `json:"icon_url,omitempty"`
 	BadgeURL                     *string `json:"badge_url,omitempty"`
-	NotificationsTTL             *int    `json:"notifications_ttl,omitempty"`
+	NotificationsTTL             *int64  `json:"notifications_ttl,omitempty"`
 	NotificationsRequireInteract *bool   `json:"notifications_require_interaction,omitempty"`
 	NotificationsSilent          *bool   `json:"notifications_silent,omitempty"`
 }
@@ -34,7 +34,7 @@ type ProjectUpdateParams struct {
 	Website                      *string `json:"website,omitempty"`
 	IconURL                      *string `json:"icon_url,omitempty"`
 	BadgeURL                     *string `json:"badge_url,omitempty"`
-	NotificationsTTL             *int    `json:"notifications_ttl,omitempty"`
+	NotificationsTTL             *int64  `json:"notifications_ttl,omitempty"`
 	NotificationsRequireInteract *bool   `json:"notifications_require_interaction,omitempty"`
 	NotificationsSilent          *bool   `json:"notifications_silent,omitempty"`
 }
